Update on paper + first 5 days training of PPO and SAC agent
Update on the paper and finished training a ppo and sac agent with 5 days of training time. The results are okay, a random agent wins less than 1% of the matches against the agent while the agent wins more than 90% of the matches, but a human player can find a couple of winning sequences after some searching.